Move alignment_check tests out
